build: migrate from Poetry/poe to uv/just
Replace Poetry (package manager) and poethepoet (task runner) with uv and
just. This simplifies the toolchain, removes the Poetry dependency, and
switches the PEP 517 build backend from poetry-core to setuptools.

- pyproject.toml: remove [tool.poetry.*], [tool.poe.*]; switch build-backend
  to setuptools.build_meta; add [dependency-groups] (PEP 735)
- setup.py: merge Rust build logic from cython_build.py; gracefully skip
  Rust build when Cargo.toml or cargo is missing (sdist installs)
- justfile: translate all poe tasks to just recipes; use uv run for
  venv-scoped commands
- MANIFEST.in: sdist includes for .pyx, .pyi, py.typed, CLAUDE.md,
  Cargo.toml, rust/ source
- CI: code-quality.yml and publish.yml use astral-sh/setup-uv@v4
- Fix CRLF line endings in scripts/wasm-opt-wrapper.sh

Deleted: cython_build.py, poetry.lock
Added: justfile, MANIFEST.in, uv.lock
