+++ .github/workflows/code-quality.yml
@@ -10,42 +10,28 @@ jobs:
   black:
     runs-on: ubuntu-latest
     name: Check code formatting with Black
-    
+
     steps:
     - uses: actions/checkout@v4
-    
-    - name: Set up Python
-      uses: actions/setup-python@v5
-      with:
-        python-version: '3.12'
-    
-    - name: Install Black
-      run: |
-        python -m pip install --upgrade pip
-        pip install black==25.11.0
-    
+
+    - name: Install uv
+      uses: astral-sh/setup-uv@v4
+
     - name: Check formatting with Black
-      run: black --check .
+      run: uv run black --check .
 
   mypy:
     runs-on: ubuntu-latest
     name: Type checking with mypy
-    
+
     steps:
     - uses: actions/checkout@v4
-    
-    - name: Set up Python
-      uses: actions/setup-python@v5
-      with:
-        python-version: '3.12'
-    
-    - name: Install dependencies
-      run: |
-        python -m pip install --upgrade pip
-        pip install mypy==1.18.2 numpy
-    
+
+    - name: Install uv
+      uses: astral-sh/setup-uv@v4
+
     - name: Run mypy
-      run: mypy .
+      run: uv run mypy .
 
   pytest:
     runs-on: ${{ matrix.os }}
@@ -61,11 +47,6 @@ jobs:
       with:
         lfs: true  # Pull test data files if using Git LFS
 
-    - name: Set up Python ${{ matrix.python-version }}
-      uses: actions/setup-python@v5
-      with:
-        python-version: ${{ matrix.python-version }}
-
     - name: Install Rust toolchain
       uses: dtolnay/rust-toolchain@stable
 
@@ -75,40 +56,29 @@ jobs:
         sudo apt-get update
         sudo apt-get install -y build-essential python3-dev g++
 
-    - name: Install Poetry (Unix)
-      if: runner.os != 'Windows'
-      run: |
-        curl -sSL https://install.python-poetry.org | python3 -
-        echo "$HOME/.local/bin" >> $GITHUB_PATH
-      shell: bash
-
-    - name: Install Poetry (Windows)
-      if: runner.os == 'Windows'
-      shell: pwsh
-      run: |
-        (Invoke-WebRequest -Uri https://install.python-poetry.org -UseBasicParsing).Content | python -
-        echo "$env:APPDATA\Python\Scripts" >> $env:GITHUB_PATH
+    - name: Install uv
+      uses: astral-sh/setup-uv@v4
 
     - name: Install dependencies
       run: |
-        poetry install
+        uv sync --python ${{ matrix.python-version }}
 
     - name: Run pytest
       run: |
-        poetry run pytest tests/ -v --tb=short
+        uv run pytest tests/ -v --tb=short
 
     - name: Run pytest (Rust backend)
       run: |
-        poetry run pytest tests/ -v --tb=short
+        uv run pytest tests/ -v --tb=short
       env:
         LIBXRK_BACKEND: rust
 
     - name: Run spec tests (full)
       run: |
-        poetry run pytest spec/tests/ -v --tb=short
+        uv run pytest spec/tests/ -v --tb=short
 
     - name: Run spec tests (Rust backend)
       run: |
-        poetry run pytest spec/tests/ -v --tb=short
+        uv run pytest spec/tests/ -v --tb=short
       env:
         LIBXRK_BACKEND: rust
