+++ .github/workflows/publish.yml
@@ -31,19 +31,14 @@ jobs:
       with:
         lfs: true
 
-    - name: Set up Python
-      uses: actions/setup-python@v5
-      with:
-        python-version: '3.12'
-
     - name: Install Rust toolchain
       uses: dtolnay/rust-toolchain@stable
 
-    - name: Install cibuildwheel
-      run: python -m pip install cibuildwheel==3.3.1
+    - name: Install uv
+      uses: astral-sh/setup-uv@v4
 
     - name: Build wheel
-      run: python -m cibuildwheel --output-dir wheelhouse
+      run: uvx cibuildwheel==3.3.1 --output-dir wheelhouse
       env:
         CIBW_BUILD: "cp${{ matrix.python-version == '3.10' && '310' || matrix.python-version == '3.11' && '311' || matrix.python-version == '3.12' && '312' || matrix.python-version == '3.13' && '313' || '314' }}-*"
         CIBW_BUILD_VERBOSITY: 1
@@ -72,18 +67,11 @@ jobs:
       with:
         lfs: true
 
-    - name: Set up Python
-      uses: actions/setup-python@v5
-      with:
-        python-version: '3.12'
-
-    - name: Install Poetry
-      run: |
-        curl -sSL https://install.python-poetry.org | python3 -
-        echo "$HOME/.local/bin" >> $GITHUB_PATH
+    - name: Install uv
+      uses: astral-sh/setup-uv@v4
 
     - name: Build sdist
-      run: poetry build --format sdist
+      run: uv build --sdist
 
     - name: Upload sdist
       uses: actions/upload-artifact@v4
@@ -107,10 +95,8 @@ jobs:
       with:
         lfs: true
 
-    - name: Set up Python ${{ matrix.python-version }}
-      uses: actions/setup-python@v5
-      with:
-        python-version: ${{ matrix.python-version }}
+    - name: Install uv
+      uses: astral-sh/setup-uv@v4
 
     - name: Download wheel artifact
       uses: actions/download-artifact@v4
@@ -118,24 +104,26 @@ jobs:
         name: wheel-${{ matrix.os }}-${{ matrix.python-version }}
         path: dist/
 
-    - name: Install wheel and test dependencies (Unix)
+    - name: Create venv and install wheel (Unix)
       if: runner.os != 'Windows'
       run: |
+        uv venv .venv --python ${{ matrix.python-version }}
         WHEEL=$(ls dist/*.whl | head -1)
-        pip install "${WHEEL}[test]"
+        uv pip install --python .venv "${WHEEL}[test]"
 
-    - name: Install wheel and test dependencies (Windows)
+    - name: Create venv and install wheel (Windows)
       if: runner.os == 'Windows'
       shell: pwsh
       run: |
+        uv venv .venv --python ${{ matrix.python-version }}
         $wheel = (Get-ChildItem dist/*.whl | Select-Object -First 1).FullName
-        pip install "${wheel}[test]"
+        uv pip install --python .venv "${wheel}[test]"
 
     - name: Run tests
-      run: pytest tests/ -v --tb=short
+      run: uv run --python .venv pytest tests/ -v --tb=short
 
     - name: Run tests (Rust backend)
-      run: pytest tests/ -v --tb=short
+      run: uv run --python .venv pytest tests/ -v --tb=short
       env:
         LIBXRK_BACKEND: rust
